+++ packages/explorer-ui/components/TransactionTable/TableHeader.tsx
@@ -2,8 +2,9 @@ export function TableHeader({ headers }) {
   return (
     <thead className="">
       <tr>
-        {headers.map((header) => (
+        {headers.map((header, index) => (
           <th
+            key={index}
             scope="col"
             className="px-2 py-2 text-left text-md font-bold text-white "
           >